refact: move servicerequestmodal to its own file
